# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Argus OFAC SDN Screening API",
    "description": "High-performance, low-latency OFAC Specially Designated Nationals (SDN) fuzzy search and compliance screening API.",
    "version": "1.0.0",
    "contact": {
      "name": "Argus API Support"
    }
  },
  "servers": [
    {
      "url": "https://argus-api-zeta.vercel.app",
      "description": "Production (Vercel)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Healthcheck & Readiness Endpoint",
        "description": "Returns operational liveness status (200 OK) or store readiness failure (503 Service Unavailable).",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "API is healthy and SDN database is fully loaded into memory.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          },
          "503": {
            "description": "API is degraded because the SDN database is not loaded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stats": {
      "get": {
        "summary": "Dataset Statistics",
        "description": "Returns comprehensive record counts, publication dates, source metadata, and memory metrics.",
        "operationId": "getStats",
        "responses": {
          "200": {
            "description": "Dataset stats retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/screen": {
      "get": {
        "summary": "Screen Entity (GET)",
        "description": "Screen an individual or entity against the OFAC SDN database via query parameters.",
        "operationId": "screenEntityGet",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "description": "Full or partial name to screen.",
            "schema": {
              "type": "string",
              "example": "Viktor Bout"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "description": "Filter by entity category.",
            "schema": {
              "type": "string",
              "enum": ["Individual", "Entity", "Vessel", "Aircraft", "all"],
              "default": "all"
            }
          },
          {
            "name": "threshold",
            "in": "query",
            "required": false,
            "description": "Minimum match similarity score (0 to 100).",
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 100,
              "default": 70
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum records to return.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "dob",
            "in": "query",
            "required": false,
            "description": "Date of birth string to boost match confidence.",
            "schema": {
              "type": "string",
              "example": "1967-01-13"
            }
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "description": "Country name or code filter/boost.",
            "schema": {
              "type": "string",
              "example": "Russia"
            }
          },
          {
            "name": "idNumber",
            "in": "query",
            "required": false,
            "description": "ID or passport number for exact identification boost.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Screening completed successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreeningResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter supplied or malformed query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal screening execution error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "SDN database not loaded yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Screen Entity (POST)",
        "description": "Screen an individual or entity against the OFAC SDN database via JSON body payload.",
        "operationId": "screenEntityPost",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreeningRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Screening completed successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreeningResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or malformed JSON body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal screening execution error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "SDN database not loaded yet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sdn/{uid}": {
      "get": {
        "summary": "Get SDN Record by UID",
        "description": "Retrieve full primary details for a specific SDN entry by numeric UID.",
        "operationId": "getSdnRecordByUid",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "description": "Unique numeric identifier of the primary SDN entry.",
            "schema": {
              "type": "integer",
              "example": 8279
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SDN record found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["record"],
                  "properties": {
                    "record": {
                      "$ref": "#/components/schemas/SdnRecord"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid non-numeric UID parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "SDN record not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "required": ["status", "uptimeSeconds", "timestamp", "recordCount"],
        "properties": {
          "status": {
            "type": "string",
            "enum": ["ok", "degraded"],
            "example": "ok"
          },
          "uptimeSeconds": {
            "type": "number",
            "example": 142.5
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "recordCount": {
            "type": "integer",
            "example": 19391
          }
        }
      },
      "StatsResponse": {
        "type": "object",
        "required": ["isLoaded", "loadedAt", "recordCount", "metadata", "memoryUsageMb"],
        "properties": {
          "isLoaded": {
            "type": "boolean",
            "example": true
          },
          "loadedAt": {
            "type": "string",
            "format": "date-time"
          },
          "recordCount": {
            "type": "integer",
            "example": 19391
          },
          "metadata": {
            "type": "object",
            "properties": {
              "publishDate": {
                "type": "string",
                "nullable": true
              },
              "recordCount": {
                "type": "integer"
              },
              "totalAkas": {
                "type": "integer"
              },
              "totalAddresses": {
                "type": "integer"
              },
              "totalIds": {
                "type": "integer"
              },
              "generatedAt": {
                "type": "string",
                "format": "date-time"
              },
              "sourceUrl": {
                "type": "string"
              }
            }
          },
          "memoryUsageMb": {
            "type": "object",
            "properties": {
              "rss": {
                "type": "string"
              },
              "heapTotal": {
                "type": "string"
              },
              "heapUsed": {
                "type": "string"
              }
            }
          }
        }
      },
      "ScreeningRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Full or partial name to screen.",
            "example": "Viktor Bout"
          },
          "type": {
            "type": "string",
            "enum": ["Individual", "Entity", "Vessel", "Aircraft", "all"],
            "default": "all"
          },
          "threshold": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "default": 70
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100,
            "default": 20
          },
          "dob": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "idNumber": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ScreeningQuery": {
        "type": "object",
        "required": ["name", "type", "threshold", "limit"],
        "properties": {
          "name": {
            "type": "string",
            "example": "Viktor Bout"
          },
          "type": {
            "type": "string",
            "example": "all"
          },
          "threshold": {
            "type": "number",
            "example": 70
          },
          "limit": {
            "type": "integer",
            "example": 20
          },
          "dob": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "idNumber": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "MatchedAlias": {
        "type": "object",
        "nullable": true,
        "required": ["fullName", "type"],
        "properties": {
          "fullName": {
            "type": "string",
            "example": "AERO-CARIBBEAN"
          },
          "type": {
            "type": "string",
            "example": "a.k.a."
          }
        }
      },
      "SecondaryMatches": {
        "type": "object",
        "required": ["dobMatched", "countryMatched", "idMatched"],
        "properties": {
          "dobMatched": {
            "type": "boolean",
            "example": false
          },
          "countryMatched": {
            "type": "boolean",
            "example": false
          },
          "idMatched": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "ScreeningMatch": {
        "type": "object",
        "required": ["uid", "sdnName", "sdnType", "programs", "score", "matchedName", "matchType", "matchedAlias", "secondaryMatches", "record"],
        "properties": {
          "uid": {
            "type": "integer",
            "example": 8279
          },
          "sdnName": {
            "type": "string",
            "example": "Viktor Anatolijevitch BOUT"
          },
          "sdnType": {
            "type": "string",
            "example": "Individual"
          },
          "programs": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["DRCONGO"]
          },
          "score": {
            "type": "number",
            "example": 100.0
          },
          "matchedName": {
            "type": "string",
            "example": "Viktor Anatolijevitch BOUT"
          },
          "matchType": {
            "type": "string",
            "enum": ["primary", "aka"],
            "example": "primary"
          },
          "matchedAlias": {
            "$ref": "#/components/schemas/MatchedAlias"
          },
          "secondaryMatches": {
            "$ref": "#/components/schemas/SecondaryMatches"
          },
          "record": {
            "$ref": "#/components/schemas/SdnRecord"
          }
        }
      },
      "ScreeningResponse": {
        "type": "object",
        "required": ["query", "totalMatches", "returnedMatches", "executionTimeMs", "results"],
        "properties": {
          "query": {
            "$ref": "#/components/schemas/ScreeningQuery"
          },
          "totalMatches": {
            "type": "integer",
            "example": 1
          },
          "returnedMatches": {
            "type": "integer",
            "example": 1
          },
          "executionTimeMs": {
            "type": "number",
            "example": 12.4
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScreeningMatch"
            }
          }
        }
      },
      "AkaEntry": {
        "type": "object",
        "description": "Alias (AKA) entry with its own sub-entry UID. Not directly queryable via /api/sdn/{uid} — use the parent SDN record's UID.",
        "properties": {
          "uid": {
            "type": "integer",
            "nullable": true,
            "example": 8280
          },
          "type": {
            "type": "string",
            "example": "a.k.a."
          },
          "category": {
            "type": "string",
            "example": "strong"
          },
          "firstName": {
            "type": "string",
            "nullable": true,
            "example": "Viktor"
          },
          "lastName": {
            "type": "string",
            "nullable": true,
            "example": "BUT"
          },
          "fullName": {
            "type": "string",
            "example": "Viktor BUT"
          },
          "normalizedName": {
            "type": "string",
            "example": "VIKTOR BUT"
          }
        }
      },
      "AddressEntry": {
        "type": "object",
        "description": "Address entry with its own sub-entry UID. Not directly queryable via /api/sdn/{uid} — use the parent SDN record's UID.",
        "properties": {
          "uid": {
            "type": "integer",
            "nullable": true,
            "example": 25
          },
          "address1": {
            "type": "string",
            "nullable": true,
            "example": "Building 7, Calle 23"
          },
          "address2": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "address3": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "city": {
            "type": "string",
            "nullable": true,
            "example": "Havana"
          },
          "stateOrProvince": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "postalCode": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "country": {
            "type": "string",
            "nullable": true,
            "example": "Cuba"
          }
        }
      },
      "IdEntry": {
        "type": "object",
        "description": "Identification document entry with its own sub-entry UID. Not directly queryable via /api/sdn/{uid} — use the parent SDN record's UID.",
        "properties": {
          "uid": {
            "type": "integer",
            "nullable": true,
            "example": 26
          },
          "idType": {
            "type": "string",
            "nullable": true,
            "example": "Tax ID No."
          },
          "idNumber": {
            "type": "string",
            "nullable": true,
            "example": "CU12345678"
          },
          "idCountry": {
            "type": "string",
            "nullable": true,
            "example": "Cuba"
          },
          "issueDate": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "expirationDate": {
            "type": "string",
            "nullable": true,
            "example": null
          }
        }
      },
      "SdnRecord": {
        "type": "object",
        "required": ["uid", "sdnType", "fullName", "normalizedName", "programs", "akas", "addresses", "ids", "dobs", "pobs"],
        "properties": {
          "uid": {
            "type": "integer",
            "example": 8279
          },
          "sdnType": {
            "type": "string",
            "example": "Individual"
          },
          "firstName": {
            "type": "string",
            "nullable": true,
            "example": "Viktor Anatolijevitch"
          },
          "lastName": {
            "type": "string",
            "nullable": true,
            "example": "BOUT"
          },
          "fullName": {
            "type": "string",
            "example": "Viktor Anatolijevitch BOUT"
          },
          "normalizedName": {
            "type": "string",
            "example": "VIKTOR ANATOLIJEVITCH BOUT"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "programs": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["DRCONGO"]
          },
          "akas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AkaEntry"
            }
          },
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddressEntry"
            }
          },
          "ids": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IdEntry"
            }
          },
          "dobs": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["13 Jan 1967"]
          },
          "pobs": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Dushanbe, Tajikistan"]
          },
          "remarks": {
            "type": "string",
            "nullable": true,
            "example": "Relating to arms trafficking."
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error", "code"],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error identifier.",
            "enum": [
              "INVALID_UID",
              "MALFORMED_JSON",
              "MISSING_NAME",
              "INVALID_THRESHOLD",
              "INVALID_TYPE",
              "INVALID_LIMIT",
              "RECORD_NOT_FOUND",
              "NOT_FOUND",
              "UNAUTHORIZED",
              "NOT_IMPLEMENTED",
              "VALIDATION_ERROR",
              "INTERNAL_ERROR"
            ],
            "example": "INVALID_UID"
          },
          "error": {
            "type": "string",
            "example": "Invalid UID parameter. Must be a numeric identifier."
          },
          "uid": {
            "type": "integer",
            "nullable": true,
            "example": 12345,
            "description": "Returned when UID is requested."
          },
          "details": {
            "type": "object",
            "nullable": true,
            "description": "Optional validation error details."
          },
          "example": {
            "type": "string",
            "nullable": true,
            "example": "/api/screen?name=Aerocaribbean&threshold=70"
          },
          "message": {
            "type": "string",
            "nullable": true,
            "description": "Detailed error message string if available."
          }
        }
      }
    },
    "securitySchemes": {
      "RapidApiSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RapidAPI-Proxy-Secret",
        "description": "RapidAPI proxy secret header for request authentication."
      }
    }
  },
  "security": [
    {
      "RapidApiSecret": []
    }
  ]
}